+++ projects/client/src/routes/movies/anticipated/+page.svelte
@@ -2,7 +2,7 @@
   import TraktPage from "$lib/components/layout/TraktPage.svelte";
   import * as m from "$lib/features/i18n/messages";
 
-  import AnticipatedPaginatedList from "$lib/sections/lists/AnticipatedPaginatedList.svelte";
+  import AnticipatedPaginatedList from "$lib/sections/lists/anticipated/AnticipatedPaginatedList.svelte";
   import { DEFAULT_SHARE_MOVIE_COVER } from "$lib/utils/constants";
 </script>
